move convenience inits to dydxViews, clean up
